# swagger.yaml
openapi: 3.0.3
info:
  title: INTVoicekit API
  description: |
    Backend API for INTVoicekit - a voice assistant platform built on Express.js with MongoDB.
    This API acts as a proxy layer between clients and the external Vyom LiveKit API,
    managing local state for users, assistants, SIP trunks, tools, inbound configurations,
    and analytics.

    ## Authentication
    Every endpoint except `POST /api/auth/signup` and `POST /api/auth/login` requires the user's
    upstream LiveKit key, sent as `Authorization: Bearer <api_key>`. A missing or unknown key is
    a `401`. Obtain the key from signup or login. Identity comes from that key — a `user_id` in
    the request body or query string is ignored, and the key is forwarded to the external Vyom
    API on the user's behalf.

    ## Dual ID Resolution
    Many endpoints accept either a local MongoDB `_id` or an external service ID
    (e.g., `external_assistant_id`, `external_trunk_id`).

    ## External API Proxy
    Most service methods proxy requests to `https://api-livekit-vyom.indusnettechnologies.com/*`,
    forwarding the user's Bearer token. External responses are typically passed through directly.

    ## Assistant Runtime Modes
    `assistant_mode` selects the session shape. It is `pipeline` by default.
    - `pipeline` (half-cascade): a realtime model emits text, an external TTS provider speaks it.
      Vendor is `openai` only — `gemini` is rejected with `400` (see below). Requires
...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
paths:
...
  /api/auth/signup:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Creates a new user account with organization details.
      security: [] # public — this is how a caller obtains the bearer key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_name, org_name, user_email, password]
              properties:
                user_name:
                  type: string
                  example: "john_doe"
                org_name:
                  type: string
                  example: "Acme Corp"
                user_email:
                  type: string
                  format: email
                  example: "john@acme.com"
                password:
                  type: string
                  format: password
                  example: "securePass123"
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "User registered successfully"
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "665f1a2b3c4d5e6f7a8b9c0d"
                      user_name:
                        type: string
                        example: "john_doe"
                      user_email:
                        type: string
                        example: "john@acme.com"
                      api_key:
                        type: string
                        nullable: true
                        example: null
        '400':
          description: Validation error or user already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Login with credentials
      description: Authenticates a user by username and password.
      security: [] # public — the password is the credential
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_name, password]
              properties:
                user_name:
                  type: string
                  example: "john_doe"
                password:
                  type: string
                  format: password
                  example: "securePass123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Login successful"
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                        example: "665f1a2b3c4d5e6f7a8b9c0d"
                      user_name:
                        type: string
                        example: "john_doe"
                      user_email:
                        type: string
                        example: "john@acme.com"
                      api_key:
                        type: string
                        example: "vyom_livekit_xxxxxxxxxxxxxxxx"
        '400':
          description: Missing credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== ASSISTANT ====================
...
  /api/call/outbound:
    post:
      tags: [Call]
      summary: Initiate an outbound call
      description: Initiates an outbound call using an assistant and SIP trunk.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assistant_id, trunk_id, to_number]
              properties:
                assistant_id:
                  type: string
                  description: Local _id or external_assistant_id
                  example: "665f1a2b3c4d5e6f7a8b9c0d"
                trunk_id:
                  type: string
                  description: Local _id or external_trunk_id
                  example: "665f1a2b3c4d5e6f7a8b9c0d"
                to_number:
                  type: string
                  example: "+1234567890"
                metadata:
                  type: object
                  description: Optional metadata passed through to the external call API
                  example: {"campaign_id": "promo_2024", "source": "web"}
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Call initiated (external API response)
          content:
            application/json:
              schema:
                type: object
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/call/queue/{queue_id}:
    get:
      tags: [Call]
      summary: Get outbound queue status
      description: >
        Dispatch state of a queued outbound call. `POST /api/call/outbound` and
        `POST /api/passthrough-call/passthrough-outbound` both return a `queue_id`; this is
        how you resolve it.

        `dispatched` means only that the handoff to the telephony provider succeeded — the
        live call outcome (`answered`, `busy`, `no_answer`, ...) arrives through the end-call
        webhook or the assistant call logs, not here.
      parameters:
        - name: queue_id
          in: path
          required: true
          schema:
            type: string
          description: The queue_id returned when the call was triggered.
          example: "8b7df5ea0fdc497ea4f44bd31954a387"
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Queue status retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: object
                    properties:
                      queue_id:
                        type: string
                      status:
                        type: string
                        enum: [pending, dispatching, dispatched, failed]
                        description: >
                          `pending` waiting for the dispatcher; `dispatching` creating the
                          room/provider call; `dispatched` handed off successfully;
                          `failed` permanently failed after retries.
                      to_number:
                        type: string
                      call_service:
                        type: string
                        enum: [twilio, exotel]
                      queued_at:
                        type: string
                        format: date-time
                      dispatched_at:
                        type: string
                        format: date-time
                        nullable: true
                      retry_count:
                        type: integer
                      last_error:
                        type: string
                        nullable: true
        '400':
          description: Invalid request parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Queue item not found for this user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== INTEGRATION ====================
...
  /api/audio/upload:
    post:
      tags: [Audio]
      summary: Upload an audio file
      description: |
        Uploads an audio file for use with voice assistants.
        Uses multipart/form-data with a file field.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [audio_name, transcript, file]
              properties:
                audio_name:
                  type: string
                  example: "Welcome Greeting"
                transcript:
                  type: string
                  example: "Hello, welcome to our service!"
                file:
                  type: string
                  format: binary
                  description: Audio file (binary)
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Audio uploaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      audio_id:
                        type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/audio/list:
    get:
      tags: [Audio]
      summary: List audio files
      description: Retrieves a paginated list of uploaded audio files.
      parameters:
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Audio list retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/audio/{audio_id}:
    get:
      tags: [Audio]
      summary: Get audio details
      description: Retrieves details for a specific audio file.
      parameters:
        - name: audio_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Audio details retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      tags: [Audio]
      summary: Delete an audio file
      description: Deletes a specific audio file.
      parameters:
        - name: audio_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '401':
          description: Missing or invalid bearer API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '200':
          description: Audio deleted
          content:
            application/json:
              schema:
                type: object
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ==================== DOCS ====================
...
components:
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: "Something went wrong"
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The user's upstream LiveKit `api_key`, obtained from `POST /api/auth/signup` or
        `POST /api/auth/login`, sent as `Authorization: Bearer <api_key>`.

# Applied to every operation; /api/auth/signup and /api/auth/login override it with `security: []`.
security:
  - bearerAuth: []